- added SmokeEffects.h and LightEffects.h to TileEngine.vcproj - increased the INT8 & UINT8 bAge/ubDuration variables in SmokeEffects.cpp and LightEffects.cpp to a good old "int".
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@192 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Weapons.h
@@ -299,7 +299,7 @@ typedef struct
 	UINT8		ubVolatility;		// maximum chance of accidental explosion
 	UINT8		ubAnimationID;	// Animation enum to use
 	UINT32		uiIndex;
-	UINT8		ubDuration;	
+	int			ubDuration;	
 	UINT8		ubStartRadius;	
 } EXPLOSIVETYPE;
